pyproject.toml
[tool.poetry]
name = "northwind_postgresql_docker"
version = "0.1.0"
description = ""
authors = ["fsimoes81 <fabio.simoes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.1"
SQLAlchemy = "^2.0.13"
psycopg2-binary = "^2.9.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
